requirements.txt
discord.py==2.3.2
jishaku==2.5.2
aiosqlite==0.20.0